flask_app/requirements.txt
google-generativeai==0.8.3
Markdown==3.7
Flask==3.0.3
google-cloud-firestore
google-cloud-storage
google-cloud-secret-manager